[frontend] fix again
