+++ opencti-platform/opencti-front/src/private/components/workspaces/dashboards/DashboardAuditsViz.tsx
@@ -14,6 +14,7 @@ import AuditsTreeMap from '@components/common/audits/AuditsTreeMap';
 import AuditsWordCloud from '@components/common/audits/AuditsWordCloud';
 import { computerRelativeDate, dayStartDate, formatDate } from '../../../../utils/Time';
 import type { Widget } from '../../../../utils/widget/widget';
+import { useRemoveIdAndIncorrectKeysFromFilterGroupObject } from '../../../../utils/filters/filtersUtils';
 
 interface DashboardAuditsVizProps {
   widget: Widget
@@ -38,6 +39,19 @@ const DashboardAuditsViz = ({
     ? formatDate(dayStartDate(null, false))
     : config.endDate;
 
+  let mainEntityTypes = ['Stix-Core-Object'];
+  if (widget.perspective === 'relationships') {
+    mainEntityTypes = ['stix-core-relationship', 'stix-sighting-relationship'];
+  } else if (widget.perspective === 'audits') {
+    mainEntityTypes = ['History'];
+  }
+  const dataSelection = widget.dataSelection.map((data) => ({
+    ...data,
+    filters: useRemoveIdAndIncorrectKeysFromFilterGroupObject(data.filters, mainEntityTypes),
+    dynamicFrom: useRemoveIdAndIncorrectKeysFromFilterGroupObject(data.dynamicFrom, ['Stix-Core-Object']),
+    dynamicTo: useRemoveIdAndIncorrectKeysFromFilterGroupObject(data.dynamicTo, ['Stix-Core-Object']),
+  }));
+
   switch (widget.type) {
     case 'number':
       return (
@@ -45,7 +59,7 @@ const DashboardAuditsViz = ({
           variant="inLine"
           endDate={endDate}
           startDate={startDate}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
         />
@@ -56,7 +70,7 @@ const DashboardAuditsViz = ({
           variant="inLine"
           endDate={endDate}
           startDate={startDate}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
         />
@@ -67,7 +81,7 @@ const DashboardAuditsViz = ({
           variant="inLine"
           endDate={endDate}
           startDate={startDate}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
         />
@@ -80,7 +94,7 @@ const DashboardAuditsViz = ({
           startDate={startDate}
           isReadOnly={isReadonly}
           withExportPopover={true}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
         />
@@ -93,7 +107,7 @@ const DashboardAuditsViz = ({
           startDate={startDate}
           isReadOnly={isReadonly}
           withExportPopover={true}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
         />
@@ -106,7 +120,7 @@ const DashboardAuditsViz = ({
           startDate={startDate}
           isReadOnly={isReadonly}
           withExportPopover={true}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
         />
@@ -119,7 +133,7 @@ const DashboardAuditsViz = ({
           startDate={startDate}
           isReadOnly={isReadonly}
           withExportPopover={true}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
         />
@@ -132,7 +146,7 @@ const DashboardAuditsViz = ({
           startDate={startDate}
           isReadOnly={isReadonly}
           withExportPopover={true}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
         />
@@ -145,7 +159,7 @@ const DashboardAuditsViz = ({
           startDate={startDate}
           isReadOnly={isReadonly}
           withExportPopover={true}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
         />
@@ -158,7 +172,7 @@ const DashboardAuditsViz = ({
           startDate={startDate}
           isReadOnly={isReadonly}
           withExportPopover={true}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
         />
@@ -170,7 +184,7 @@ const DashboardAuditsViz = ({
           endDate={endDate}
           startDate={startDate}
           withExportPopover={true}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
         />
@@ -183,7 +197,7 @@ const DashboardAuditsViz = ({
           startDate={startDate}
           isReadOnly={isReadonly}
           withExportPopover={true}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
         />
@@ -194,7 +208,7 @@ const DashboardAuditsViz = ({
           variant="inLine"
           endDate={endDate}
           startDate={startDate}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
         />
