+++ opencti-platform/opencti-front/src/private/components/workspaces/dashboards/DashboardRelationshipsViz.tsx
@@ -17,6 +17,7 @@ import StixRelationshipsMap from '@components/common/stix_relationships/StixRela
 import StixRelationshipsWordCloud from '@components/common/stix_relationships/StixRelationshipsWordCloud';
 import { computerRelativeDate, dayStartDate, formatDate } from '../../../../utils/Time';
 import type { Widget } from '../../../../utils/widget/widget';
+import { useRemoveIdAndIncorrectKeysFromFilterGroupObject } from '../../../../utils/filters/filtersUtils';
 
 interface DashboardRelationshipsVizProps {
   widget: Widget
@@ -41,6 +42,19 @@ const DashboardRelationshipsViz = ({
     ? formatDate(dayStartDate(null, false))
     : config.endDate;
 
+  let mainEntityTypes = ['Stix-Core-Object'];
+  if (widget.perspective === 'relationships') {
+    mainEntityTypes = ['stix-core-relationship', 'stix-sighting-relationship'];
+  } else if (widget.perspective === 'audits') {
+    mainEntityTypes = ['History'];
+  }
+  const dataSelection = widget.dataSelection.map((data) => ({
+    ...data,
+    filters: useRemoveIdAndIncorrectKeysFromFilterGroupObject(data.filters, mainEntityTypes),
+    dynamicFrom: useRemoveIdAndIncorrectKeysFromFilterGroupObject(data.dynamicFrom, ['Stix-Core-Object']),
+    dynamicTo: useRemoveIdAndIncorrectKeysFromFilterGroupObject(data.dynamicTo, ['Stix-Core-Object']),
+  }));
+
   switch (widget.type) {
     case 'number':
       return (
@@ -48,7 +62,7 @@ const DashboardRelationshipsViz = ({
           variant="inLine"
           endDate={endDate}
           startDate={startDate}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
         />
@@ -60,7 +74,7 @@ const DashboardRelationshipsViz = ({
           endDate={endDate}
           startDate={startDate}
           widgetId={widget.id}
-          dataSelection={widget.dataSelection} // dynamicFrom and dynamicTo TODO
+          dataSelection={dataSelection} // dynamicFrom and dynamicTo TODO
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
         />
@@ -71,7 +85,7 @@ const DashboardRelationshipsViz = ({
           variant="inLine"
           endDate={endDate}
           startDate={startDate}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
           title={undefined} // because calling js component in ts
@@ -87,7 +101,7 @@ const DashboardRelationshipsViz = ({
           startDate={startDate}
           isReadOnly={isReadonly}
           withExportPopover={true}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
         />
@@ -100,7 +114,7 @@ const DashboardRelationshipsViz = ({
           startDate={startDate}
           isReadOnly={isReadonly}
           withExportPopover={true}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
         />
@@ -113,7 +127,7 @@ const DashboardRelationshipsViz = ({
           startDate={startDate}
           isReadOnly={isReadonly}
           withExportPopover={true}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
           title={undefined} // because calling js component in ts
@@ -125,7 +139,7 @@ const DashboardRelationshipsViz = ({
           variant="inLine"
           endDate={endDate}
           startDate={startDate}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
         />
@@ -138,7 +152,7 @@ const DashboardRelationshipsViz = ({
           startDate={startDate}
           isReadOnly={isReadonly}
           withExportPopover={true}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
           title={undefined} // because calling js component in ts
@@ -147,8 +161,8 @@ const DashboardRelationshipsViz = ({
       );
     case 'horizontal-bar':
       if (
-        widget.dataSelection.length > 1
-        && widget.dataSelection[0].attribute === 'internal_id'
+        dataSelection.length > 1
+        && dataSelection[0].attribute === 'internal_id'
       ) {
         return (
           <StixRelationshipsMultiHorizontalBars
@@ -157,7 +171,7 @@ const DashboardRelationshipsViz = ({
             startDate={startDate}
             isReadOnly={isReadonly}
             withExportPopover={true}
-            dataSelection={widget.dataSelection}
+            dataSelection={dataSelection}
             parameters={widget.parameters as object} // because calling js component in ts
             height={undefined} // because calling js component in ts
             title={undefined} // because calling js component in ts
@@ -172,7 +186,7 @@ const DashboardRelationshipsViz = ({
           startDate={startDate}
           isReadOnly={isReadonly}
           withExportPopover={true}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
           title={undefined} // because calling js component in ts
@@ -190,7 +204,7 @@ const DashboardRelationshipsViz = ({
           startDate={startDate}
           isReadOnly={isReadonly}
           withExportPopover={true}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
           title={undefined} // because calling js component in ts
@@ -205,7 +219,7 @@ const DashboardRelationshipsViz = ({
           startDate={startDate}
           isReadOnly={isReadonly}
           withExportPopover={true}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
           title={undefined} // because calling js component in ts
@@ -220,7 +234,7 @@ const DashboardRelationshipsViz = ({
           startDate={startDate}
           isReadOnly={isReadonly}
           withExportPopover={true}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
         />
@@ -233,7 +247,7 @@ const DashboardRelationshipsViz = ({
           startDate={startDate}
           isReadOnly={isReadonly}
           withExportPopover={true}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
           title={undefined} // because calling js component in ts
@@ -246,7 +260,7 @@ const DashboardRelationshipsViz = ({
           variant="inLine"
           endDate={endDate}
           startDate={startDate}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
           title={undefined} // because calling js component in ts
@@ -259,7 +273,7 @@ const DashboardRelationshipsViz = ({
           variant="inLine"
           endDate={endDate}
           startDate={startDate}
-          dataSelection={widget.dataSelection}
+          dataSelection={dataSelection}
           parameters={widget.parameters as object} // because calling js component in ts
           height={undefined} // because calling js component in ts
           title={undefined} // because calling js component in ts
